interface/requirements.txt
Flask==2.2.5
paho_mqtt==1.6.1
python-dotenv==1.0.1
Requests==2.32.3
